[ci] Use CI's credentials for image pushing instead of gcr-push (#14301)
CI owns the internal container registry, so there's no reason to have an
additional service account that CI impersonates in order to push images.
This removes uses of `registry-push-credentials` in favor of using the
CI identity that already exists in the batch jobs that interact with the
container registry. I've marked this as `full-deploy` because you want
to roll out this change before a subsequent PR that I will make to
delete the `gcr-push` identity.