[hailtop] in WithoutSemaphore, optionally do not acquire the semaphore when raising
Supposes there are 10,000 tasks competing for the semaphore. If any one of them rasises
an exception while the `WithoutSempahore` context manager is active, then the probability
that the `WithoutSemaphore.__aexit__` wins the semaphore is 1/10,000 aka unlikely. In practice,
this means we need to process through most of the tasks to get the first exception. If,
for a real world example, there is a permission issue, you will wait many minutes for all
10,000 tasks to fail, all with the same exception, and then you will finally receive
the first exception plus a *litany* of log messages about other tasks which were
not properly cleaned up (due to the implementation of `bounded_gather2_raise_exceptions`).