[ci] Retry more for auth copy paste logins (#14353)
I don't have great evidence but 15 seconds might not be very much if we
get big preemptions in k8s and our auth pods need to be redeployed on
new nodes. I've been seeing more timeouts in Azure CI in this step and
thought we should retry a little longer. If this issue persists I'll
devote some time to digging deeper on it.