[query] Add AlleleType Enum
This exposes functionality previously present in the private
`hail.expr.functions._allele_ints` and
`hail.expr.functions._num_allele_type`, used to avoid strings in the
query when determining allele type.

Implement AlleleType as a python [IntEnum].

Replace `_num_allele_type` with the now public `numeric_allele_type`.

As a note to developers of hail, it is _unlikely_ that the enum values
of AlleleType will change, but they are documented as though they could.

CHANGELOG: Exposed previously internal _num_allele_type as
           numeric_allele_type and deprecated it. Add new AlleleType
           enumeration for users to be able to easily use the values
           returned by numeric_allele_type.

[IntEnum]: https://docs.python.org/3/library/enum.html#enum.IntEnum