[batch] Dont grab an X lock on the batches table in attempts after update (#14379)
By joining against the `batches` row in the insert `FOR UPDATE`, these
queries try to grab an exclusive lock on the row in the `batches` table.
This is a problem if the transaction already has a shared lock on that
row, like it does in `mark_job_complete`. In that case any two
concurrent MJCs from the same batch will deadlock.