Prevent iteration of Job and ResourceGroup objects
These classes define __getitem__() so their resources can be accessed
as if via a dict. Python will use __getitem__(0),... to implement
iteration through the object; define __iter__() to prevent this.

Especially for ResourceGroup, it may be worth defining iteration for
these classes in future. But at the moment __getitem__-based iteration
fails, so this ensures it fails with a clear TypeError message.