[query/vds] Don't error if combiner is finished and output exists (#14397)
We want to error if the output exists since the combiner will overwrite
the output path. However we don't want to do this if the combiner is
done, since if the user is rerunning a script with the combiner in the
middle of it, they are generally doing some post-processing on the
combiner's output and we (and the users) would probably rather just skip
the combiner if possible.