[batch] Close the aiohttp client session used by the gcp WorkerAPI (#14418)
When workers shut down, we get warning logs about unclosed client
sessions, as seen in #14261. While it can be difficult to derive the
source of the client session, I think it's the one held by the
`GCPWorkerAPI`. I've added an exit stack and added the session's close
method to it.

I also made a small change to `Worker`. I find it can be difficult to
determine whether or not a particular class should close a client
session because it's not always clear who owns it. Without a clear way
to communicate this in python, I think we should just never transfer
ownership of a `ClientSession` and always assume that if an object's
constructor takes a client session, it should be assumed a borrow and
not close the session when the object is closed. As such, I moved the
call to `client_session()` into the `Worker` constructor so it's clear
that the worker owns the session.