[query] reorganize Binds.scala (#14404)
This PR refactors the `Binds.scala` file, which encodes the binding
structure of the IR.

The binding structure controls what variable bindings are in scope in
each location of the IR. The encoding specifies, for each node `x` and
each child `c` of `x`, how does the environment (the set of bindings
which are in scope, and their types) of 'x' differ from the environment
of `c`. For instance, `c`'s environment might add a new bindings (e.g.
in the body of a `StreamMap`), or it might drop all bindings, isolating
the child, so that it cannot refer to anything bound outside of `x`
(e.g. all relational nodes).

The binding structure is complicated by the separation of the
environment into "eval", "agg", "scan", and "relational" scopes. The
child `c`'s environment might add new bindings in any of these scopes,
and it might modify the parent's environment in other ways, like
replacing the "eval" scope with the "agg" scope.

In `Binds.scala`, the encoding of this relationship between parent's and
child's environments was split across many methods, separately encoding
bindings added to each of the scopes, and a prior modification to the
parent scope. This made it difficult to understand the binding behavior
of any particular node. And for nodes with more complicated binding
behavior, these separate encodings became confusingly entangled.

In this PR, `Binds.scala` is rewritten so that the binding behavior of a
node is specified in a single place. Instead of computing a list of new
bindings, it simply takes the parent's environment and returns the
child's.

Some use cases do require knowing what new bindings are added by a node
in a child's environment. To support this, I made the
`BindingEnvironment` interface a trait, and added another implementor
`SegregatedBindingEnv`, which wraps two separate `BindingEnvironment`s,
and puts all new bindings into the second. This satisfies the invariant
that for any function `f(GenericBindingEnv): GenericBindingEnv` which
modifies an environment using only the `GenericBindingEnv` interface,
and any `env: BindingEnv[T]`,
```
f(env) == f(SegregatedBindingEnv(env)).unified
```
i.e. the segregation doesn't have any effect, other than keeping track
of which bindings are new.

Now that all binding structure for a node is encoded in one place, a
future change could move this encoding to live in the node class
directly.