[query] Unify Let and AggLet into new Block node (#14402)
~Stacked on #14404~

This set of changes started out as refactoring and trying to better
understand `extract.scala`. But the bigger change ended up being getting
rid of AggLet in favor of allowing bindings in `Let` to be agg/scan
bindings.

So each binding in a `Let` now has a binding scope, encoded using the
pre-existing `Scope` enum. I also renamed `Let` to `Block`. While I
think that's a better name for what `Let` has become, and what I hope it
will further become in the future, I initially tried to keep the name
`Let`. I changed it so that I could keep `Let.apply` the way it is,
specialized to bindings all in the eval scope. We can change all the
uses of `Let.apply` seperately, I didn't want to pollute this already
large pr with that.