[release] build release wheel with correct `deploy.yaml`
In a89d64a, I modified `build.yaml` to release the wheel we had already
built and tested. Unbeknownst to me was that we rebuild the wheel with a
different version of `hail/python/hailtop/hailctl/deploy.yaml` and
releasing the version used for testing borked `hailctl dataproc`
commands.

To fix this, we'll rebuild the wheel but use the `jar` we've already
built and tested. This is safe to do as far as I know because we don't
bundle any information into the jar that depends on the make flag
`DEPLOY_REMOTE`.

Fixes: #14452