[compiler] rewrite FreeVariables to not duplicate binding structure (#14451)
Before we can simplify the binding structure, we need to stop
duplicating it all over the place. This PR rewrites `FreeVariables` so
that it no longer needs special logic for particular nodes, hard coding
binding structure (redundantly). To do this, it takes advantage of the
new `Bindings`, which operates on a `GenericBindingEnv` interface. It
adds a new implementation of this interface specifically for computing
free variables, then simply does a generic traversal of the IR using
this custom binging environment.

While I find the new implementation far simpler and more obviously
correct than the old, I do expect it to further simplify once I'm able
to start modifying the core binding structure.