[qob] Use job groups for driver job and worker stages (#14432)
In QoB, the Query Driver and all Query Worker jobs reside in the same
batch. Currently, the Query Driver submits worker jobs and then waits
for them to finish before collecting their results from GCS. The way it
waits is by polling the status of the batch and waiting for the number
of completed jobs to reach `n_total_jobs - 1` (to account for itself).
This is both awkward and prevents a couple potentially valuable pieces
of functionality: you cannot run multiple concurrent Query Drivers, and
you cannot take advantage of Batch's `cancel_after_n_failures`
functionality because you do not want to cancel the Query Driver itself.

The introduction of job groups addresses both these problems, as the
Query Driver can create a job group for the stage of worker jobs and
then await its completion. This PR makes a job group for the query
driver and then the query driver creates a nested job group per stage of
workers that it creates.

Utilizing `cancel_after_n_failures` to fail a stage early is future
work.