[terra] Fix terra on azure helm chart and make targets (#14450)
For Hail Batch on Terra Azure, the production artifact is Helm chart
containing the necessary kubernetes resources to run a Hail Batch
deployment in a Terra k8s cluster. This deployment contains slightly
modified containers of the batch front-end, batch driver and a mysql
database. This chart is currently built manually using the targets in
`batch/terra-chart/Makefile`. As this process is not currently
automatically tested, it's very prone to bit rot. This PR is an
amalgamation of fixes that I needed to make to get `main` to build in
the current Terra. A non-exhaustive list of the changes are:

- After changing from gradle to mill, the some Dockerfiles and make
targets needed to change to account for the new location of the JAR.
- I removed some redundancy in invocations of `docker build` by relying
on the generic targets that we now have in the top level Makefile.
- Terra changed how they handle identity management for the kubernetes
deployment, from `aadpodidentity` to `workloadIdentity`. This changes
the chart to work with their new inputs they provide.

Ultimately, terra should have a CI system that we can push charts to and
receive feedback on whether it passed our test suite in their test
environment.