Fix region pool being unset in CompiledEncoder
Before now, encoders never needed to allocate. Now, depending on the
type of the encoded value, EStructOfArrays may need to allocate to build
intermidiate arrays. In worker executed code, this is never an issue, as
the partition region is always set, but in driver executed code, we
usually don't call `resultWithIndex` for the class builder, and instead
call `result`, so in cases where the compiled code expects a region
pool, we have to add one ourselves.

Accomplish this by making Encoder.writeRegionValue take a region, and
use that region to set a region pool on the now
FunctionWithPartitionRegion implementing encoder class.