[pip] Update orjson and transitive dependencies (#14471)
`orjson` 3.9.15 fixed the rare segfault that we saw in `3.9.11`. Besides
just updating to latest patch and minor versions:

- Removed a redundant requirement of `orjson` in `gear/requirements.txt`
-- it inherits `orjson` from hailtop
- Bokeh `3.4` made a breaking change w.r.t. the `circle` method on
figures. I have restricted the bounds for `bokeh` to avoid this breaking
change but will follow up with a PR that changes our usage of bokeh to
follow the deprecation/upgrade advice and undo the bounds restriction