[pip] Use `uv` for generating lockfiles (#14503)
[uv](https://github.com/astral-sh/uv) is a new package resolver by the
same folks who make `ruff`. It's boasted for being really fast, which
honestly it is, but mostly it's appealing to me because they support
generating lockfiles for alternative platforms and python versions than
the system you run it on, which allows us to delete all this dockerizing
`pip-compile` in order to generate lockfiles for linux.

It's a really green project, so I'm open to pushback on incorporating
it, but it seemed like a worthwhile simplification. I also quite like
that it allows for additional strategies in generating lockfiles. By
default, it behaves as would be expected, where it locks packages to the
highest version within the acceptable bounds. But you can also configure
it to generate the *lowest* acceptable pins, so we could actually verify
whether the lower bounds that we have in our requirements files are
actually acceptable or not.