[batch] Remove unused cost_from_msec_mcpu code (#14511)
This function is an unused relic of the first version of our billing
system, that just computed the cost of a slot on a Batch VM per mcpu
milliseconds based on some hard-coded prices. We long ago migrated to a
system where we track GCP SKUs and price changes through their API and
properly itemize usage by SKU in the database.