[compiler] consolidate BindingEnv interface (#14475)
As a step towards coralling the specification of the binding structure
of the IR into one place, this rewrites `Bindings` to use only a single
method of the `GenericBindingEnv` interface, `newBlock`, which therefore
captures all possibilities of how a node can modify its parent's
environment in a child. Later work refactors this to return an object
encoding this modification, instead of returning a modified environment,
which allows the caller complete flexibility in how to maintain an
environment appropriately for their use case.

This PR leaves in the old `Bindings` implementation, with an assertion
that they agree. The PR stacked above this, #14495, deletes the old
implementation. This way CI asserts that this refactoring hasn't changed
any behavior.