[compiler] declarative `Bindings` (#14496)
Refactors `Bindings` to return an object encoding the change to the
environment (any new bindings, whether the agg/scan env is promoted,
etc). This allows the deletion of `SegregatedBindingEnv`. Follow up work
will use this to replace the other specializations of
`GenericBindingEnv`, and to greatly simplify compiler passes, such as
`NormalizeNames` and `PruneDeadFields`, which currently need to
redundantly encode the binding structure of every node.