[batch] fix g2 machine memory calculation (#14498)
The memory that machines use is calculated based off of the number of
cores the machine has, with a fix ratio set of cores to Gb of RAM. The
code assumed a ratio of 3.75 cores: 1 Gb but this assumption does not
hold outside of the n1 machine family. This PR changes the functions
that calculate memory from number of cores to account for this by
passing the machine_type as the main parameter rather than the worker
type. Then, logic is added to use the appropriate ratio of 4:1 for the
g2 family of machines.

---------

Co-authored-by: Daniel Goldstein <danielgold95@gmail.com>
Co-authored-by: Sophie Parsa <parsa@wm9c6-e4e.broadinstitute.org>