[dependabot] Remove dependabot automatic PRs (#14504)
At time of writing dependabot doesn't have a great way to bulk update
dependencies across unrelated lockfiles in the repo, which often require
manual intervention because we assert that we're always using the same
package version everywhere. It's also just a lot of noise and hogs CI
time. We've decided to move to a periodic bulk-update process + using
repo Security alerts.