[query] Struct of Arrays encoder prototype (#14391)
Create an `EType`, `EStructOfArrays`, that encodes an
`array<struct{...}>` into a transposed representation (hopefully)
allowing for better data compression (either specialized per field, or
simply better patterns for a general purpose algorithm to find) and
decoding speed.

At current, this is a prototype that only supports structs where all
fields are `int32`.

Each field of the struct of arrays is it's own `EContainer`, so we have
fine grained control over the encoding of each field.