[hailtop] Dont assume exact error message match for ClientPayloadError retrying (#14545)
The treatment of `ClientPayloadError` as a sometimes transient error was
originally made in response to [an existing
issue](https://github.com/aio-libs/aiohttp/issues/4581) in aiohttp that
can cause transient errors on the client that are difficult to
distinguish from a real broken server. What's in `main` matched exactly
on the error message, but that error message has [since
changed](https://github.com/aio-libs/aiohttp/commit/dc38630b168a169139974617d75e176530c91696)
to include more information, breaking our transient error handling. This
change relaxes the requirement of the error response string to fix
transient error handling for our current version of `aiohttp`.

I wish I had a better approach. `ClientPayloadError` can also be thrown
in the case of malformed data, so I am reticent to treat it as always
transient, but we could perhaps make it a `limited_retries_error` and
avoid inspecting the error message.