[compiler] refactor types in bindings (#14517)
Slightly tweak the `Bindings` class to store a single array of (name,
type) pairs, with the various environments referring back to that array.
This simplifies passes that want to create a single mutable state for
each binding, even those that are bound in multiple environments (e.g.
eval and agg).