[batch] Use more precise terminology around instance cost, price and revenue (#14543)
The batch code makes copious use of the word "cost", which can cause
confusion when considering "whose cost?". This PR attempts to use more
precise language (definitions provided by Investopedia):

- Cost is the expense incurred for making a product
- Price is what a customer would pay for a product

Since we are the "company" in question in these definitions, we define
what the cloud bills us as the cost of an instance, and what we charge
users as the price. Note that the price is higher than the cost because
we charge a service fee. This can be a little tricky to get right across
the codebase because we as batch operators play both the role as the
company to the user and customer to the cloud provider. All of the code
changed here though pertains to batch as a company and the definitions
should match accordingly.