[compiler] add newtype for bound names in IR (#14547)
This PR introduces a new type `Name` for representing bound variables in
the IR, replacing `String`. For now, it is just an `AnyVal` wrapper
around `String`, but in the future I would like to take advantage of the
new type. For example, I'd like to:
* change equality of `Name` from string comparison to comparing object
identity with `eq`. That way `freshName` becomes just `new Name()`, with
stronger guarantees that the new name doesn't occur anywhere in the
current IR, without needing to maintain global state as we do now.
* get rid of `NormalizeNames`, instead enforcing the global uniqueness
of names as a basic invariant of the IR (typecheck could also check this
invariant)
* keep a string in the `Name`, but no longer require it to be unique.
Instead it's just a suggestion for how to show the name in printouts,
adding a uniqueifying suffix as needed. With `NormalizeNames` gone, this
would let us preserve meaningful variable names further in the lowering
pipeline.
* possibly keep other state in the `Name`, for example to allow a more
efficient implementation of environments, similar to the `mark` state on
`BaseIR`

This is obviously a large change, but there are only a few conceptual
pieces (appologies for not managing to separate these out):
* attempt to minimize the number of locations in which the `Name`
constructor is called, to make future refactorings easier
* add `freshName()`, which just wraps `genUID()`, returning a `Name`
* convert IR construction to use the convenience methods in
`ir.package`, which take scala lambdas to represent blocks with bound
variables, instead of manually creating new variable names
* replace uses of the magic constant variable names (`row`, `va`, `sa`,
`g`, `global`) with constants (`TableIR.{rowName, globalName}`,
`MatrixIR.{rowName, colName, entryName, globalName}`)
* the above changes modified the names we use for bound variables in
many places. That shouldn't matter, but it cought a couple bugs where it
did.
* `NormalizeNames` optionally allows the IR to contain free variables.
But it didn't do anything to ensure the newly generated variable names
are distinct from any contained free variables. Thus it was possible to
rename a bound variable to mistakenly capture a contained free variable.
I've fixed that.
* `SimplifySuite` compared simplified IR with the pre-constructed
expected IR, carefully controlling the `genUID` global state to make
simplify generate exactly the names expected. I've replaced that by just
comparing with the expected IR using a alpha-equivalence comparison.