[batch] Add resource_usage endpoint for jobs (#14328)
Hi! 

I know this is out of the blue, but we would like the ability to fetch
resource_usage data from an endpoint programmatically to evaluate our
job performance. I thought it might be worth suggesting this upstream to
see if it's something you'd like too :)

This PR:
1. Use the internal method to fetch the dataframes for a job
2. Transform the data frame to dictionary with `orient='split'`

And FWIW, here's how to convert it back into a dataframe:

```python
import pandas as pd

response = {} # response from Hail Batch
dataframes = {
    key: pd.DataFrame(data=values['data'], columns=values['columns'])
    for key, values in response.items()
}
```

I tested this in on a dev deploy and it worked pretty well, but happy to
add testing if you can direct me to a place to add it.

---------

Co-authored-by: Michael Franklin <illusional@users.noreply.github.com>
Co-authored-by: John Marshall <john.marshall@populationgenomics.org.au>