[compiler] simplify PruneDeadFields (#14509)
Makes some further progress on simplifying the `PruneDeadFields` pass,
with the primary goal of decoupling it from the details of the binding
structure.

The primary change is to `memoizeValueIR`. Before, it passed in only the
requested type of the node, and returned and environment containing all
free variables and their requested types. Any bound variables would then
need to be removed, and the environments of all children then merged.
This low-level manipulation of environments made it closely tied to the
binding structure, essentially redundantly encoding everything in
`Binds.scala`.

Now we pass an environment down into the children, which maps variables
to a mutable state tracking the requested type. Each `Ref` node unions
the requested type at the reference with the state in the environment.
This lets us use the general environment infrastructure.

I didn't do an assertion directly comparing the old and new
implementations, as I've done with some other pass rewrites. But
`PruneDeadFields` has pretty good test coverage.