[infra] Update GCP DB terraform to reflect backup region change (#14568)
I manually changed the Cloud SQL automated backups storage from the `us`
multi-region to `us-central1`. There's no reason to store it in a
multi-region and it's more expensive. What I didn't realize is that this
configuration is actually owned by the terraform that we have managing
the lifecycle of the database and other infra, so we need to update the
terraform to reflect the desired (and current) state.