[batch] Redesign batch front end pages (#14562)
This revamps the `batch.front_end` web pages to be a bit less cluttered
and more usable. I brought in tailwindcss instead of using our existing
sass. I mainly find it easier to work with but am fine porting it to
sass if anyone feels strongly against changing css tools. In order to
avoid overhauling all our HTML in one PR, I added a flag to the jinja
context `use_tailwind` that is currently only on for the `batch`
service. `layout.html` then uses this flag to determine which CSS and
header it should use. If this gets to a state we're happy to merge, it
should be a lot less work to bring over auth and CI and delete that
flag.