[hailtop] Prevent event loop RecursionError during highly concurrent file uploads (#14576)
We use `nest_asyncio` to allow synchronous blocking on a coroutine
execution inside an async context. For example, if a user is using the
synchronous interface of hail inside a jupyter cell (which runs an event
loop). `nest_asyncio` achieves this by patching the `asyncio` event loop
to make it reentrant, and with that there are footguns. This allows us
to do things like create 100 tasks that all concurrently invoke
`run_until_complete`, each of which will add stack frames to the event
loop stack that can pile up and trigger a cryptic `RecursionError`.

But internally we never *need* to make concurrent calls to
`run_until_complete`, and more broadly we should never have `async /
sync / async` inside hail code. This change exposes an asynchronous
`validate_file` so that asynchronous methods in `hailtop` can use it
directly instead of inserting a synchronous layer (`async_to_blocking`)
between them.